[AI] Add shared plugin foundation packages
