+++ .gitignore
@@ -33,9 +33,17 @@ packages/desktop-electron/dist
 packages/desktop-electron/loot-core
 packages/desktop-client/service-worker
 packages/plugins-service/dist
+packages/plugins-core/build
+packages/plugins-core/dist
+packages/plugins-core-sync-server/dist
+packages/sync-server-plugin-example/dist
 packages/component-library/dist
+packages/query/dist
+packages/shared-types/dist
 packages/loot-core/lib-dist
-**/.tsbuildinfo
+**/*.tsbuildinfo
+packages/*/*.zip
+packages/sync-server/server-files/plugins/*.zip
 packages/sync-server/coverage
 bundle.desktop.js
 bundle.desktop.js.map
