+++ packages/loot-core/src/types/models/account.ts
@@ -1,26 +1,4 @@
-import type { BankSyncProviders } from './bank-sync';
-
-export type AccountEntity = {
-  id: string;
-  name: string;
-  offbudget: 0 | 1;
-  closed: 0 | 1;
-  sort_order: number;
-  last_reconciled: string | null;
-  tombstone: 0 | 1;
-
-  // Sync fields
-  account_id: string | null;
-  bank: string | null;
-  bankName: string | null;
-  bankId: string | null;
-  mask: string | null; // end of bank account number
-  official_name: string | null;
-  balance_current: number | null;
-  balance_available: number | null;
-  balance_limit: number | null;
-  account_sync_source: AccountSyncSource | null;
-  last_sync: string | null;
-};
-
-export type AccountSyncSource = BankSyncProviders;
+export type {
+  AccountEntity,
+  AccountSyncSource,
+} from '@actual-app/shared-types/models/account';
