+++ packages/loot-core/src/types/models/category-group.ts
@@ -1,11 +1,1 @@
-import type { CategoryEntity } from './category';
-
-export type CategoryGroupEntity = {
-  id: string;
-  name: string;
-  is_income?: boolean;
-  sort_order?: number;
-  tombstone?: boolean;
-  hidden?: boolean;
-  categories?: CategoryEntity[];
-};
+export type { CategoryGroupEntity } from '@actual-app/shared-types/models/category-group';
