+++ packages/loot-core/src/types/models/category.ts
@@ -1,14 +1,1 @@
-import type { CategoryGroupEntity } from './category-group';
-
-export type CategoryEntity = {
-  id: string;
-  name: string;
-  is_income?: boolean;
-  group: CategoryGroupEntity['id'];
-  goal_def?: string;
-  cleanup_def?: string;
-  template_settings?: { source: 'notes' | 'ui' };
-  sort_order?: number;
-  tombstone?: boolean;
-  hidden?: boolean;
-};
+export type { CategoryEntity } from '@actual-app/shared-types/models/category';
