+++ packages/loot-core/src/types/models/payee.ts
@@ -1,10 +1,1 @@
-import type { AccountEntity } from './account';
-
-export type PayeeEntity = {
-  id: string;
-  name: string;
-  transfer_acct?: AccountEntity['id'];
-  favorite?: boolean;
-  learn_categories?: boolean;
-  tombstone?: boolean;
-};
+export type { PayeeEntity } from '@actual-app/shared-types/models/payee';
