+++ packages/shared-types/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@actual-app/shared-types",
+  "version": "0.0.1",
+  "private": true,
+  "description": "Shared type definitions for Actual Budget",
+  "license": "MIT",
+  "exports": {
+    "./models/*": "./src/models/*.ts",
+    "./modalProps": "./src/modalProps.ts",
+    "./spreadsheet": "./src/spreadsheet.ts",
+    "./util": "./src/util.ts"
+  },
+  "scripts": {
+    "build": "tsgo -b",
+    "typecheck": "tsgo -b"
+  },
+  "devDependencies": {
+    "@types/react": "^19.2.14",
+    "@typescript/native-preview": "beta",
+    "react": "19.2.4",
+    "typescript-strict-plugin": "^2.4.4"
+  }
+}
