+++ packages/shared-types/src/modalProps.ts
@@ -1,0 +1,11 @@
+import type { CSSProperties } from 'react';
+
+export type BasicModalProps = {
+  isLoading?: boolean;
+  noAnimation?: boolean;
+  style?: CSSProperties;
+  onClose?: () => void;
+  containerProps?: {
+    style?: CSSProperties;
+  };
+};
