+++ packages/shared-types/src/models/bank-sync.ts
@@ -1,0 +1,9 @@
+export const SYNC_PROVIDERS = [
+  'goCardless',
+  'simpleFin',
+  'pluggyai',
+  'enableBanking',
+] as const;
+
+export type BuiltInBankSyncProvider = (typeof SYNC_PROVIDERS)[number];
+export type BankSyncProviders = BuiltInBankSyncProvider | string;
