+++ packages/shared-types/src/models/category-views.ts
@@ -1,0 +1,7 @@
+import type { CategoryEntity } from './category';
+import type { CategoryGroupEntity } from './category-group';
+
+export type CategoryViews = {
+  grouped: CategoryGroupEntity[];
+  list: CategoryEntity[];
+};
