+++ upcoming-release-notes/8070.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [lelemm]
+---
+
+Add shared package foundations for upcoming plugin support.
